backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
google-cloud-aiplatform==1.71.0
google-generativeai==0.8.3
google-genai>=1.5.0
google-cloud-storage==2.18.2
google-cloud-firestore>=2.16.0
Pillow==10.4.0
numpy>=1.26.0
httpx>=0.28.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
scipy>=1.12.0
